Adicionado documentação swagger, logging slf4j e método toString

+++ src/main/java/com/payco/banco/model/ContaPoupanca.java
@@ -4,7 +4,7 @@ public class ContaPoupanca extends Conta{
 
 	private int aniversariopoupanca;
 
-	public ContaPoupanca(Long Id, String titular, int cpf, int numeroConta, float saldo,int aniversariopoupanca) {
+	public ContaPoupanca(Long Id, String titular, Long cpf, int numeroConta, float saldo,int aniversariopoupanca) {
 		super(Id, titular, cpf, numeroConta, saldo);
 		this.aniversariopoupanca = aniversariopoupanca;
 	}
